requirements.txt
# MD Converter — Python dependencies
# Install: pip install -r requirements.txt

PyMuPDF>=1.24.0
python-docx>=1.1.0
pywebview>=5.0
requests>=2.31.0
beautifulsoup4>=4.12.0
markdownify>=0.13.0
striprtf>=0.0.26
certifi>=2024.0.0
Pillow>=10.0.0
pytesseract>=0.3.10

# macOS native drag-and-drop (optional — app works without it)
pyobjc-core>=10.0; sys_platform == "darwin"
pyobjc-framework-Cocoa>=10.0; sys_platform == "darwin"
